--- tests/test_rules.py
@@ -1,6 +1,0 @@
-import pytest
-from wsgame import *
-
-def test_rules():
-    rules = Rules()
-
