+++ tests/test_wsgame.py
@@ -1,4 +1,4 @@
-from wsgame import Game, make_app
+from wsgame import *
 import pytest
 import urllib.parse
 import json
@@ -28,7 +28,7 @@ class Helper(object):
 
 
 
-class TestGame(testing.AsyncHTTPTestCase):
+class TestWSGame(testing.AsyncHTTPTestCase):
     def get_app(self):
         return make_app()
 
@@ -38,59 +38,77 @@ class TestGame(testing.AsyncHTTPTestCase):
         h = Helper(self)
 
         ws1 = yield from h.connect()
+        event, data = yield from h.recv(ws1)
+        assert WSGame.EVT_ADD_PLAYER == event
+        assert data == {"player": "unknown1", "players": {"unknown1": "resistance"} }
 
-        h.send(ws1, Game.EVT_JOIN, {"player": "causton"})
+        h.send(ws1, WSGame.EVT_CHANGE_NAME, {"player": "causton"})
 
         event, data = yield from h.recv(ws1)
-        assert Game.EVT_ADD_PLAYER == event
+        assert WSGame.EVT_ADD_PLAYER == event
         assert data == {"player": "causton", "players": {"causton": "resistance"} }
 
 
         ws2 = yield from h.connect()
-        h.send(ws2, Game.EVT_JOIN, {"player": "kaladin"})
+        h.send(ws2, WSGame.EVT_CHANGE_NAME, {"player": "kaladin"})
 
         event, data = yield from h.recv(ws2)
-        assert Game.EVT_ADD_PLAYER == event
+        assert WSGame.EVT_ADD_PLAYER == event
+        assert data == {"player": "unknown2", "players": {"causton": "resistance", "unknown2": "resistance"} }
+        event, data = yield from h.recv(ws2)
+        assert WSGame.EVT_ADD_PLAYER == event
         assert data == {"player": "kaladin", "players": {"causton": "resistance", "kaladin": "resistance"} }
 
         event, data = yield from h.recv(ws1)
-        assert Game.EVT_ADD_PLAYER == event
+        assert WSGame.EVT_ADD_PLAYER == event
+        assert data == {"player": "unknown2", "players": {"causton": "resistance", "unknown2": "resistance"} }
+        event, data = yield from h.recv(ws1)
+        assert WSGame.EVT_ADD_PLAYER == event
         assert data == {"player": "kaladin", "players": {"causton": "resistance", "kaladin": "resistance"} }
 
 
-        h.send(ws1, Game.EVT_ASSIGN_ROLES, {})
+        h.send(ws1, WSGame.EVT_ASSIGN_ROLES, {})
 
         event, data = yield from h.recv(ws1)
-        assert Game.EVT_ERROR == event
-        assert {'message': Game.MSG_NEED_MORE_PLAYERS } == data
+        assert WSGame.EVT_ERROR == event
+        assert {'message': bundle.fill(NOT_ENOUGH_PLAYERS) } == data
 
         socks = [ws1, ws2]
         players = {"causton": "resistance", "kaladin": "resistance"}
         for i in range(3,6):
             name = "player{}".format(i)
+            unk = "unknown{}".format(i)
             ws = yield from h.connect()
             socks.append(ws)
 
-            h.send(ws, Game.EVT_JOIN, {"player": name})
+            h.send(ws, WSGame.EVT_CHANGE_NAME, {"player": name})
+            players[unk] = 'resistance'
+
+            results = yield [s.read_message() for s in socks]
+            for res in results:
+                event, data = json.loads(res)
+                assert WSGame.EVT_ADD_PLAYER == event
+                assert data == {"player": unk, "players": players }
+
+            players.pop(unk)
             players[name] = 'resistance'
 
             results = yield [s.read_message() for s in socks]
-
-
             for res in results:
                 event, data = json.loads(res)
-                assert Game.EVT_ADD_PLAYER == event
+                assert WSGame.EVT_ADD_PLAYER == event
                 assert data == {"player": name, "players": players }
 
 
 
-        msg = [Game.EVT_ASSIGN_ROLES, {}]
+
+        msg = [WSGame.EVT_ASSIGN_ROLES, {}]
         ws1.write_message(json.dumps(msg))
         results = yield [s.read_message() for s in socks]
 
         for res in results:
             event,data = json.loads(res)
-            assert Game.EVT_ASSIGN_ROLES == event
+            assert WSGame.EVT_ASSIGN_ROLES == event
 
 
 
@@ -99,7 +117,7 @@ class TestGame(testing.AsyncHTTPTestCase):
         res = yield ws2.read_message()
         del players['causton']
         event, data = json.loads(res)
-        assert Game.EVT_REM_PLAYER == event
+        assert WSGame.EVT_REM_PLAYER == event
         #assert data == {"player": "causton", "players": players }
         assert data["player"] == "causton"
         assert "causton" not in data["players"]
